code review & bug fix round 4 phase 10
